requirements.txt
tensorflow[and-cuda]==2.15.1
tensorflow-model-optimization==0.7.5
numpy==1.26.4
pillow==12.0.0